python-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic[email]==2.5.3
bcrypt==4.1.2
PyJWT==2.8.0
stripe==8.0.0
boto3==1.34.34
python-multipart==0.0.6
mangum==0.17.0
